requirements.txt
gradio
pydantic==2.10.6 # Stable version as of now
google-genai
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
sentence-transformers
unstructured[pdf]
unstructured[docx]
unstructured[ppt]
unstructured[excel]
unstructured[xml]
torch
torchaudio
transformers
fastapi
uvicorn
numpy
python-dotenv